RefMod.jl

Functions

RefMod.amsa — Function

Calculates the reflectance using the AMSA (Advanced Modified Shadowing and Coherent Backscattering) model.

Arguments

single_scattering_albedo: The single scattering albedo.

incidence_direction: The incidence direction.

emission_direction: The emission direction.

surface_orientation: The surface orientation.

phase_function: The phase function.

roughness::Float64=0.0: The roughness.

hs::Float64=0.0: The shadowing parameter.

bs0::Float64=0.0: The shadowing parameter.

hc::Float64=0.0: The coherent backscattering parameter.

bc0::Float64=0.0: The coherent backscattering parameter.

Returns

Array{Float64}: An array representing the reflectance values.

Raises

ArgumentError: If the phase function parameters are unsupported.

RefMod.coef_a — Function

coef_a(n::Int64=15)

Calculates the coefficients 'a_n' for the Legendre polynomial series.

Arguments

n::Int64=15: The number of coefficients to calculate. Default is 15.

Returns

Array{Float64}: An array of coefficients 'a_n' for the Legendre polynomial series.

Note

Equation 27 in (Hapke, 2002).

RefMod.coef_b — Function

coef_b(b::Float64=0.21, c::Float64=0.7, n::Int=15)

Calculates the coefficients for the Hapke reflectance model Legendre polynomial expansion.

Arguments

b::Float64=0.21: The single scattering albedo.

c::Float64=0.7: The asymmetry factor.

n::Int=15: The number of coefficients to calculate.

Returns

Array{Float64}: The calculated coefficients for the Legendre polynomial expansion.

Note

Equation on page 530 in (Hapke, 2002). Default values for b and c are for the moon, taken from (Warell, 2004).

RefMod.double_henyey_greenstein — Function

Calculates the phase function for the double Henyey-Greenstein model.

Arguments:

cos_g::AbstractArray{<:AbstractFloat}: The cosine of the scattering angle.

b::Float64=0.21: The asymmetry parameter.

c::Float64=0.7: The backscatter fraction.

Returns:

AbstractArray{<:AbstractFloat}: The phase function value.

Note:

Default values for b and c are for the moon, taken from (Warell, 2004).

RefMod.function_p — Function

function_p(x::AbstractArray{<:AbstractFloat}, b::Float64=0.21, c::Float64=0.7, n::Int=15)

Calculates the P function using the Hapke reflectance model.

Arguments

x::AbstractArray{<:AbstractFloat}: The input array.

b::Float64=0.21: The single scattering albedo.

c::Float64=0.7: The asymmetry factor.

n::Int=15: The number of coefficients to calculate.

Returns

Array{Float64}: The calculated P function.

Note

Equations 23 and 24 in (Hapke, 2002). Default values for b and c are for the moon, taken from (Warell, 2004).

RefMod.h_function — Function

Calculates the Hapke function for a given set of parameters.

Arguments:

x::Union{AbstractFloat,AbstractArray{<:AbstractFloat}}: The input parameter.

w::Union{AbstractFloat,AbstractArray{<:AbstractFloat}}: The weight array.

level::Int64=1: The accuracy level of the Ambartsumian–Chandrasekhar function to calculate.

derivative::Bool=false: Whether to calculate the derivative.

Returns:

Union{AbstractFloat,AbstractArray{<:AbstractFloat}}: The calculated Hapke function value.

Raises:

ArgumentError: If an invalid level is provided.

RefMod.h_function_1 — Method

Calculates the Ambartsumian–Chandrasekhar function level 1 for a given set of parameters.

Arguments:

x::Union{AbstractFloat,AbstractArray{<:AbstractFloat}}: The input parameter.

w::Union{AbstractFloat,AbstractArray{<:AbstractFloat}}: The weight array.

Returns:

Union{AbstractFloat,AbstractArray{<:AbstractFloat}}: The calculated Ambartsumian–Chandrasekhar function level 1 value.

Note:

Equation 2 in (Hapke, 2002).

RefMod.h_function_2 — Method

Calculates the Ambartsumian–Chandrasekhar function level 2 for a given set of parameters.

Arguments

x::Union{AbstractFloat,AbstractArray{<:AbstractFloat}}: The input parameter.

w::Union{AbstractFloat,AbstractArray{<:AbstractFloat}}: The weight array.

Returns:

Union{AbstractFloat,AbstractArray{<:AbstractFloat}}: The calculated Ambartsumian–Chandrasekhar function level 2 value.

Note:

Equation 13 in (Hapke, 2002).

RefMod.microscopic_roughness — Method

Calculates the microscopic roughness factor for the Hapke reflectance model.

Arguments

roughness::Float64: The roughness parameter.

incidence_direction::AbstractArray{Float64}: Array of incidence directions.

emission_direction::AbstractArray{Float64}: Array of emission directions.

surface_orientation::AbstractArray{Float64}: Array of surface orientations.

Returns

s::AbstractArray{Float64}: The microscopic roughness factor.

cos_i_s::AbstractArray{Float64}: The modified incidence-normal cosine value $\mu_i'$.

cos_e_s::AbstractArray{Float64}: The modified emission-normal cosine value $\mu'$.

Note

prime-zero terms: equations 48 and 49 in (Hapke, 1984).

$i < e$: equations 46 and 47 in (Hapke, 1984).

$i \ge e$: equations 50 and 51 in (Hapke, 1984).

RefMod.value_p — Function

value_p(b::Float64=0.21, c::Float64=0.7, n::Int=15)

Calculates the value of the P function.

Arguments

b::Float64=0.21: The single scattering albedo.

c::Float64=0.7: The asymmetry factor.

n::Int=15: The number of coefficients to calculate.

Returns

Float64: The calculated P value.

Note

Equations 25 in (Hapke, 2002). Default values for b and c are for the moon, taken from (Warell, 2004).

References

Hapke, B. (1984). Bidirectional reflectance spectroscopy: 3. Correction for macroscopic roughness. Icarus 59, 41–59.

Hapke, B. (2002). Bidirectional Reflectance Spectroscopy: 5. The Coherent Backscatter Opposition Effect and Anisotropic Scattering. Icarus 157, 523–534.

Warell, J. (2004). Properties of the Hermean regolith: IV. Photometric parameters of Mercury and the Moon contrasted with Hapke modelling. Icarus 167, 271–286.